implement scraping of the onchain hashi state (#74)
This patch introduces the OnchainState struct, the sole mechanism for accessing onchain state inside of the hashi service.

For right now this only does a single scraping of the state on startup and doesn't track any updates. As we introduce ways for onchain state to be updated, we'll introduce a monitor/watching system to do fine-grained updates of the service's view of the onchain state, as well as the ability to periodically re-scrape sections of the onchain state.

This will evolve to also support allowing subsystems in the hashi service to be able to subscribe to updates so that they will be notified of events or updates relevant to that subsystem.